Added AstVisitor, *ngFor parsing
Supported more blocking features of current parser.

With the exception of foo=“{{bar}}”, it should be possible to parse almost all templates in the wild today that work in Angular 2.2.0+

There is also enough done where it could start to be integrated with pkg/angular2, though we’d still need to write a new directive resolver.