Accept any amount of event suffixes, ie keyup.ctrl.shift.enter etc (#57)
* Accept any amount of event suffixes, ie keyup.ctrl.shift.enter etc

Decided to call the "reductions", for no real good reason. Basically,
they filter down the events, but "filter" sounded to me like something
to linear -- ie, keyup.ctrl.a should be a filter of keyup.a, but that's
not the case. For whatever reason, I felt like reductions didn't imply
this while still being a synonym of filter.

We *could* enforce a, say, 6 limit based on the current modifiers etc
that are allowed for keyup, but that seems arbitrary.

Alternatively, we could try to parse keyup/down modifiers and keys
directly in the ast, rather than leaving it as an array of strings.
That's probably the best long term option.

At the very least, this will allow us to, in the plugin, support
keyup.foo.bar.baz rather than report it as an error. Later on we can
try to validate it in either place, once we've solved the false errors.